refactor: remove backend from storage to backend directory
